requirements.txt
streamlit
PIL
google.generativeai
datetime    